Added appropriate macros for everything the headerizer demands in imcc.y
git-svn-id: https://svn.parrot.org/parrot/trunk@26186 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ compilers/imcc/sets.c
@@ -72,8 +72,8 @@ PARROT_CANNOT_RETURN_NULL
 Set*
 set_make_full(int length)
 {
-    Set * const s   = set_make(length);
-    const int bytes = NUM_BYTES(length);
+    Set * const s      = set_make(length);
+    const size_t bytes = NUM_BYTES(length);
 
     if (bytes)
         memset(s->bmp, 0xff, bytes);
